libertas/sdio: simplify suspend/resume.
The current suspend/resume code is wrong.
Returning -ENOSYS does not cause the device to be powered
off, it causes the suspend to fail.

If the card is currently active (interface is 'up'), we do
need to power-off the card, and then power-on again at resume.

This doesn't handle WoW, and is currently just a hack.

Signed-off-by: NeilBrown <neilb@suse.de>