net-next: mediatek: add support for mt7623
Add support for SoCs from the mt7623 family. With this generation of SoCs
we have seen a change from MIPS to ARM cores. Due to the fact that ARM SoC
tend to be a bit more complex than MIPS, we need support for power domains.

The code should also run on MT2701, however this has not been tested yet on
device. To my understanding only changes to the devicetree are required as
irqs and reset gpios have changed. The DMA issues have also been solved on
this silicon allowing us to run on pure QDMA without having to bringup the
PDMA engine. However, in order for QDMA to work, we need to bringup the 3rd
DMA engine called FQ_DMA. This just needs to be setup and is more scratch
memory used internally by the SoC. The reason is that this SoC has many new
advanced features such as internal SFQ QoS and HW_LRO. for these to work,
the SoC needs extra DMA memory to be able to reorganize the packets
internally. I have not had time yet to look at these new features in
detail.

Signed-off-by: John Crispin <blogic@openwrt.org>
Signed-off-by: Felix Fietkau <nbd@openwrt.org>
Signed-off-by: Michael Lee <igvtee@gmail.com>