mtd:m25p80.c: support reboot on GNUBEE-PC1
This chip boots in 3-byte addressing mode, but needs 4-byte
to access all 32K.
So gnubee-pc1 bios sets 4-byte mode.
However the SoC cannot boot when in 4-byte mode as it
assumes 3-byte mode.  So we need to reset to 3-byte mode
for reboot.
The hardware doesn't do this, so we add a reboot handler
to the software.  On a clean reboot, at least, the flash
is reset so that the CPU can boot from it.

This was taken from a patch attributed to
   Nikolay Martynov <mar.kolya@gmail.com>
but that patch did several other things, and the
changelog comment didn't mention this change, so
I'm not certain who really wrote it.

Signed-off-by: NeilBrown <neil@brown.name>