MIPS: ralink: remove ralink_halt()
ralink_halt() does nothing that machine_halt()
doesn't already do, so it adds no value.

It actually causes incorrect behaviour due to the
"unreachable()" at the end.  This tell the compiler that the
end of the function will never be reached, which isn't true.
The compiler responds by not adding a 'return' instruction,
so control simply moves on to whatever bytes come afterwards
in memory.  In my tested, that was the ralink_restart()
function.  This means that an attempt to 'halt' the machine
would actually cause a reboot.

So remove ralink_halt() so that a 'halt' really does halt.

Signed-off-by: NeilBrown <neil@brown.name>