Clean up kernel compilation warnings
Several files have static functions that are always defined but are
only invoked within #ifdef blocks that are preprocessed away in the
gnubee configuration, resulting in unused-function warnings. Put
these in #ifdef blocks so they are elided when they are unused.
Similarly, one static variable declaration is moved to where it's used
so that it's not flagged as an unused variable.

Some static functions and variables really are unused, so I deleted
them.

Two functions are reported as having misleading indentation;
indentation is fixed here.