lib: don't depend on linux headers being installed.
gen_crc64table requires linux include files to be installed
in /usr/include/linux.  This is a new requrement so hosts that
could previously build the kernel, now cannot.

gen_crc64table makes this requirement by including <linux/swab.h>,
but nothing from that header is actaully used.

So remove the #include, so that the linux headers no longer
need to be installed.

Fixes: feba04fd2cf8 ("lib: add crc64 calculation routines")
Signed-off-by: NeilBrown <neil@brown.name>