lustre: lustre: don't use l_wait_event() for poll loops.
This patch is from OpenSFS lustre.  We already had most of
the changes in linux-lustre, but to keep the two code bases
as close as practicable, change the loops here to match
the code that landed in OpenSFS.

Original comment message was:

When polling without any usable wait queue, it is clearest
to have an explicit poll loop.
So don't use l_wait_event() in these two cases, but
use a while loop with ssleep(1);

WC-bug-id: https://jira.whamcloud.com/browse/LU-10467
Lustre-commit: d0ca764a1a91 ("LU-10467 lustre: don't use l_wait_event() for poll loops.")
Signed-off-by: Mr NeilBrown <neilb@suse.com>
Reviewed-on: https://review.whamcloud.com/35968
Reviewed-by: James Simmons <jsimmons@infradead.org>
Reviewed-by: Shaun Tancheff <stancheff@cray.com>
Reviewed-by: Petros Koutoupis <pkoutoupis@cray.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Signed-off-by: NeilBrown <neilb@suse.de>