lnet: always put a page list into struct lnet_libmd.
'struct lnet_libmd' is only created in lnet_md_build().
It can be given a list of pages or a virtual address.
In the latter case, the memory will eventually be
split into a list of pages.
It is cleaner to split it into a list of pages early
so that all lower levels only need to handle one type: a page list.

Signed-off-by: NeilBrown <neilb@suse.com>