lustre: always pass struct lnet_md by reference.
Both LNetMDAttach and LNetMDBind expected a struct lnet_md
to be passed by value.  This requires coping the data structure
onto the stack, which is a waste of stack space and brings
no value.

So change the to expect a reference, and declare it 'const' to
be sure it doesn't get changed.

Signed-off-by: NeilBrown <neilb@suse.com>