lnet: discard ksnn_lock
This lock in 'struct ksock_net' is being taken in places where it
isn't needed, so it is worth cleaning up.

The lock is only needed:
 - to ensure concurrent updates to ksnn_npeers don't trace, and
 - to ensure that no more peers are added after the net is shutdown.

The first is best achieved using atomic_t.
The second is more easily achieved by replacing the ksnn_shutdown
flag with a large negative bias on ksnn_npeers, and using
atomic_inc_unless_negative().

Signed-off-by: NeilBrown <neilb@suse.com>