lnet: prepare to make lnet_lnd const.
Preferred practice is for structs containing function
pointers to be 'const'.  Such structs are generally tempting
attack vectors, and making them const allows linux to place
them in read-only memory, this reducing the attack surface.

'struct lnet_lnd' is mostly function pointers, but contains
two writable fields - a list_head and a refcount.

Rather than keeping registered lnds in a linked-list, we can place them
in an array  indexed by type - type numbers are at most 15
so this is not a burden.
The array which contains a pointer to the lnd, can also contain
the refcount.

With these changes, no part of an lnet_lnd is ever modified.

Signed-off-by: NeilBrown <neilb@suse.com>