lnet: centralise management of lnd_refcount
lnet_find_lnd_by_type() will return the type with refcount
  incremented.
__lnet_find_lnd_by_type() will find it without locking or
  incrementing.
lnet_put_lnd() will drop the refcount.

This ensures refcount manipulation is consistently done under
the lock, and simplifies the code in lnet_startup_lndnet().

lnet_start_lndni did take the ln_lnd_mutex, but that is not necessary,
so it is removed.

Signed-off-by: NeilBrown <neilb@suse.com>