lnet: o2iblnd: simplify kiblnd_connd()
It is not immediately obvious from reading kiblnd_connd()
what it waits for.  It makes explicit use of a wait_entry and
schedule_timeout(), rather than using the simple wait_event*
function.

By introducing a wait_event_interruptible_timeout() call
we can simplify the code, discarding the "dropped_lock" flag and only
holding the spinlock when we actually need it.

As a slight safety improvement, replace
  if (timeout <= 0)
with
  while (timeout <= 0)

If timeout were ever <0, and it was passed to
schedule_timeout(), that would trigger a warning.

Signed-off-by: NeilBrown <neilb@suse.com>