lnet: o2iblnd: simplify kiblnd_scheduler()
This patch moves the locking around so that the lock is taken when it
is needed, rather than dropped when it is not wanted.
This doesn't change the number of times the lock is taken, it just
moves them in the code.

This allows the waiting to be simplified into a
wait_event_interruptible() call, and allows did_something to be
discarded.

Signed-off-by: NeilBrown <neilb@suse.com>