lnet: o2iblnd: inline kiblnd_disconnect_conn()
kiblnd_disconnect_conn() is a small function with only one call-site.
It contains two assertions that become trivially obvious if
the code is explicit at that call-site.

So discard the function and replace the call with the code from the
function, excluding those assertions.
As a result, we can discard a field from struct kib_data.

Signed-off-by: NeilBrown <neilb@suse.com>