lnet: introduce wait_var_event_warning.
LNet has a recurring pattern of waiting for some variable to
reach a particular value, and generating a warning every second that
it hasn't.  I many cases the warning has a higher priority if the
wait has been for a power-of-2 seconds.

This patch embodies that pattern in a new macro
  wait_var_event_warning()

and uses wake_up_var() to cause the wait to complete as soon as
possible.

Signed-off-by: NeilBrown <neilb@suse.de>