lnet: use kernel_accept() rather than lnet_sock_accept()
The only advantage of lnet_sock_accept() over kernel_accept()
appears to be that it can be interrupted by calling
  sk->sk_state_change()

kernel_accept() can be interrupted by sending a signal
to the process, and as it is only used from a dedicated process
we can easily use a signal to interrupt.

So discard lnet_sock_accept(), use kernel_accept(), and send
a signal.

Signed-off-by: NeilBrown <neilb@suse.com>