lnet: socklnd: use jiffies in ksocknal_reaper()
Timeouts in ksocknal_reaper are typically about one second.
Calculating the timeout to a resolution of one second is
likely to produce rounding errors.  While these errors
are not likely to be of much consequence, it does mean that the code
doesn't do quite what it looks like it is meant to do.  i.e. it is
harder to understand.

The mostly likely consequence is that it will sometime
check_peer_timeouts for 2 chunks at a time, rather than always for
one.

So convert it (back) to using jiffies.

Signed-off-by: NeilBrown <neilb@suse.com>