lnet: socklnd: use wait_event_idle_timeout in ksocknal_reaper
ksocknal_reaper ends with code which performs much the same
function as wait_event_idle_timeout().  So change to use that.

There is a smaller difference - the current code doesn't loop.
However every place that wakes up the waitqueue also makes a change so
that the test will fail, so the loop in wait_event_idle_timeout() will
not repeat more than once.

Signed-off-by: NeilBrown <neilb@suse.com>