lnet: eq: discard struct lnet_handle_eq
The Portals API uses a cookie 'handle' to identify an EQ
This is appropriate for a user-space API for objects maintained by
the kernel, but it brings no value when the API client and
implementation are both in the kernel, as is the case with Lustre and
LNet.

Instead of using a 'handle', a pointer to the 'struct lnet_eq` can be
used.
This object is not reference counted and is always freed correctly, so
they can be no case where the cookie becomes invalid while it is still
held.

So use 'struct lnet_eq *' directly instead of having indirection
through a 'struct lnet_handle_eq'.
Also:
 - have LNetEQAttach() return the pointer, using ERR_PTR() to return
   errors.
 - discard ln_eq_containers and don't store the me there-in.
   This means we don't free any eq that have not already been freed,
   but all eq that are allocated are properly freed, so that is not
   a problem.

Signed-off-by: NeilBrown <neilb@suse.com>