lnet: don't use LNetEQPoll for ping replies.
lnet_ping() is the only user of LNetEQPoll.  All other
event queues register a callback and don't make use of
queuing.
The queuing provided by lib-eq isn't really needed in
the kernel where callbacks are so easy.
So as a step towards simplifying lib-eq, change lnet_ping()
to register a callback which uses a completion to notify
when the ping is finally complete.

Signed-off-by: NeilBrown <neilb@suse.com>