LNet: ptlrpc: use percpu refcount to track outstanding requests.
ptlrpc needs to wait for outstanding rpc requests to complete before
ptlrpc_ni_fini() finishes.
It currently does this using a refcount in the event-queue, but that
refcount is otherwise unused, and it will shortly be removed.

So a a percpu refcount that explicitly tracks these requests, and wait
for it with a completion.

Also move the declaration of ptlrpc_eq to ptlrpc_internal.h which is a
more appropriate location.

Signed-off-by: NeilBrown <neilb@suse.com>