lnet: store event handler directly in lnet_md
lnet_md current holds a pointer to an "event-queue"
which stores a callback function.  This pointer is only
used to find and call that function.  The function never
changes.

Instead, we can store the callback function pointer directly in the
MD.  This avoids the need to hold a counted reference on the
event-queue for the lifetime of the md.
Nothing monitors the refcount (any more) so this doesn't hurt.

Signed-off-by: NeilBrown <neilb@suse.com>