lnet: clarify initial refcount on lpni_refcount
An lnet_peer_ni is *always* added to the hash table.
So rather than incrementing the refcount at the point,
initialize it to one.

This is more normal usage, and allows the atomic_inc when the ni is
added to ln_remote_peer_ni_list to be the expected
lnet_peer_ni_addref_locked().

Signed-off-by: NeilBrown <neilb@suse.de>