lnet: remove locking protection ln_testproto_compat
lnet_net_lock(LNET_LOCK_EX) is a heavy-weight lock that is not
necessary here.  The bits in this field are only set rarely - via an
ioctl - and the pattern for reading and clearing them exactly matches
test_and_clear_bit().
So change the field to "unsigned long" (so test_and_clear_bit() can be
used), and use test_and_clear_bit(), discarding all other locking.

Signed-off-by: NeilBrown <neilb@suse.com>