lustre: replace lprocfs_climp_check()
The usage pattern for lprocfs_climp_check() is clumsy.
It must be paired with up_read(&obd->u.cli.cl_sem), but the
two calls look very different so the connection isn't obvious.

Replace it with a structure macro/statement:

 with_obd_cl_sem(rc, obd, imp) {
     statements;
 }

statements are protected by the semaphore and only run in imp can be
set to a non-NULL pointer.
rc can be changed by the statements, and should be returned afterwards
as it may have been set to -ENODEV.

Signed-off-by: NeilBrown <neilb@suse.com>