lustre: utils: update wirecheck for new values
Update the wirecheck.c file to handle changes that were made directly
in the wiretest.c files.  This allows the wiretrst.c file to be
regenerated properly without losing checks for new constants, and
fixes issues with some #defines that were changed to named enums.

Code under CONFIG_FS_POSIX_ACL in utils/wirecheck.c does not build in
userspace if the <linux/posix_acl_xattr.h> header is not available.
It should be enough that we are checking this in ptlrpc/wirecheck.c
since struct posix_acl_xattr_entry and posix_acl_xattr_header come
from the kernel anyway, so a userspace check is mostly redundant.

Fixes: 1fc602326303 ("lustre: utils: lfs rmfid")
Fixes: cbd398f591a9 ("lustre: ptlrpc: Add increasing XIDs CONNECT2 flag")
Fixes: c52da9b97ee0 ("lustre: introduce CONFIG_LUSTRE_FS_POSIX_ACL")
Fixes: 90a3f96db913 ("lustre: mdt: improve IBITS lock definitions")
Fixes: a203ed722315 ("lustre: osd: Add nonrotational flag to statfs")
WC-bug-id: https://jira.whamcloud.com/browse/LU-12937
Lustre-commit: bc2e23e1cd80 ("LU-12937 utils: update wirecheck for new values")
Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-on: https://review.whamcloud.com/36706
Reviewed-by: Artem Blagodarenko <c17828@cray.com>
Reviewed-by: Arshad Hussain <arshad.super@gmail.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Signed-off-by: NeilBrown <neilb@suse.de>