lnet: use kfree_rcu() to free lnet_ping_target.
lnet_ping_target is changed occasionally, so we currently use
lnet_net_lock_current() to protect it while taking a reference.

A more efficient way to take a safe reference is to use
use atomic_inc_not_zero() while in an RCU protected region, and
to free the old ping buffer using kfree_rcu().  This requires a simple
retry loop to handle the rare case when we raced with the
ping buffer being replaced.

We don't want to require kfree_rcu for freeing *all* ping buffers,
only those used as lnet_ping_target - so add a flag to the
lnet_ping_buffer to record the particular usage of each buffer.

As we now access the new ping_target without any locking, we need
rcu_assign_pointer() and rcu_dereference() to ensure we don't access
stale data.

Signed-off-by: NeilBrown <neilb@suse.com>